Revert "Feat: Přidat do database service absence queue na ukládání dat o skenu"
