+++ services/database/kysely/types.ts
@@ -1,25 +1,19 @@
-import type { ColumnType } from "kysely";
+import type { ColumnType } from "npm:kysely";
 export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
   ? ColumnType<S, I | undefined, U>
   : ColumnType<T, T | undefined, T>;
 export type Timestamp = ColumnType<Date, Date | string, Date | string>;
-
 export type Absence = {
-    userId: string;
-    id: Generated<string>;
-    lesson: number;
-    date: Timestamp;
-};
-export type AbsenceScan = {
-    id: Generated<string>;
-    absences: unknown;
+  userId: string;
+  id: Generated<string>;
+  lesson: number;
+  date: Timestamp;
 };
 export type User = {
-    id: Generated<string>;
-    name: string;
+  id: Generated<string>;
+  name: string;
 };
 export type DB = {
-    Absence: Absence;
-    AbsenceScan: AbsenceScan;
-    User: User;
+  Absence: Absence;
+  User: User;
 };
