+++ services/database/prisma/schema.prisma
@@ -25,8 +25,3 @@ model Absence {
   lesson Int
   date   DateTime
 }
-
-model AbsenceScan {
-  id       String @id @default(dbgenerated("gen_random_uuid()"))
-  absences Json
-}
