+++ services/database/src/main.ts
@@ -1,12 +1,9 @@
-import { RabbitMQ } from "./rabbitmq.ts";
-import { UserRequestQueue } from "./queues/user-requeust-queue.ts";
+import { RabbitMQ, UserRequestQueue } from "./rabbitmq.ts";
 import { port } from "./constants.ts";
 import { userRouter } from "./routes/users/users.ts";
 import { absenceRouter } from "./routes/absences.ts";
 import { Application, oakCors } from "./deps.ts";
 import { Meta, log } from "./logger.ts";
-import { AbsenceQueueController } from "./queues/absence-queue.ts";
-import { ScanQueue } from "./queues/scan-queue.ts";
 
 // Oak stuff
 export const app = new Application();
@@ -43,22 +40,8 @@ app.listen({ port });
 
 // RabbitMQ stuff
 const client = await RabbitMQ.init();
-const channel = await client.createChannel();
 // User request queue
+const channel = await client.createChannel();
 const queue = await client.createQueue(channel, "user_request_queue");
 const userRequestQueue = new UserRequestQueue(client, channel, queue);
 userRequestQueue.consumeFromQueue();
-
-// ScanQueue
-const scanQueue = await client.createQueue(channel, "scan:shift");
-const scanQueueController = new ScanQueue(client, channel, scanQueue);
-
-// Absence queue
-const absenceQueue = await client.createQueue(channel, "absence_queue");
-const absenceQueueController = new AbsenceQueueController(
-  client,
-  channel,
-  absenceQueue,
-  scanQueueController
-);
-absenceQueueController.consumeFromQueue();
