--- services/database/src/utils/uint8Array-to-json.ts
@@ -1,9 +1,0 @@
-export const uint8ArrayToJson: <T>(arr: Uint8Array) => T = (arr) => {
-  const decoder = new TextDecoder();
-  const decoded = decoder.decode(arr);
-  try {
-    return JSON.parse(decoded);
-  } catch {
-    return [];
-  }
-};
