+++ services/gateway/handlers/scan/consume.go
@@ -71,6 +71,12 @@ func ScanWs(c *websocket.Conn) {
 
 			c.WriteJSON(data)
 
+			if data["status"] == "DONE" {
+				c.Close()
+				ch.Close()
+				break
+			}
+
 		}
 	}()
